requirements/runtime.txt
chumpy
dataclasses; python_version == '3.6'
json_tricks
matplotlib
munkres
numpy
opencv-python
pillow
scipy
torchvision
xtcocotools>=1.6
future
tensorboard
mmpose
